+++ src/components/molecules/Input/index.ts
@@ -1,0 +1,1 @@
+export * from './Input';
